refactor code a bit
