+++ arcgis_map_sdk/lib/arcgis_map_sdk.dart
@@ -4,6 +4,5 @@ library arcgis_map;
 export 'package:arcgis_map_sdk/src/arcgis_location_display.dart';
 export 'package:arcgis_map_sdk/src/arcgis_map_controller.dart';
 export 'package:arcgis_map_sdk/src/arcgis_map_sdk.dart';
-export 'package:arcgis_map_sdk/src/model/auto_pan_mode.dart';
 export 'package:arcgis_map_sdk/src/model/map_status.dart';
 export 'package:arcgis_map_sdk_platform_interface/arcgis_map_sdk_platform_interface.dart';
