+++ arcgis_map_sdk/lib/src/arcgis_location_display.dart
@@ -1,4 +1,4 @@
-import 'package:arcgis_map_sdk/src/model/auto_pan_mode.dart';
+import 'package:arcgis_map_sdk_platform_interface/src/types/auto_pan_mode.dart';
 import 'package:arcgis_map_sdk_platform_interface/arcgis_map_sdk_platform_interface.dart';
 
 /// The use case for manual location displays is relevant when the application
