+++ arcgis_map_sdk_platform_interface/lib/arcgis_map_sdk_platform_interface.dart
@@ -10,6 +10,7 @@ export 'package:arcgis_map_sdk_platform_interface/src/models/user_position.dart'
 export 'package:arcgis_map_sdk_platform_interface/src/models/view_position.dart';
 export 'package:arcgis_map_sdk_platform_interface/src/types/arcgis_map_options.dart';
 export 'package:arcgis_map_sdk_platform_interface/src/types/attributes.dart';
+export 'package:arcgis_map_sdk_platform_interface/src/types/auto_pan_mode.dart';
 export 'package:arcgis_map_sdk_platform_interface/src/types/bounding_box.dart';
 export 'package:arcgis_map_sdk_platform_interface/src/types/default_widget.dart';
 export 'package:arcgis_map_sdk_platform_interface/src/types/elevation_mode.dart';
