+++ arcgis_map_sdk_platform_interface/lib/src/types/auto_pan_mode.dart
